Fix some warnings.

+++ lib/database/src/sb/SbVec3d.cpp
@@ -127,7 +127,7 @@
  double
  SbVec3d::length() const
  {
-     return sqrtf(vec[0] * vec[0] + vec[1] * vec[1] + vec[2] * vec[2]);
+     return sqrt(vec[0] * vec[0] + vec[1] * vec[1] + vec[2] * vec[2]);
  }
 
  //
